requirements.txt
playwright
playwright
pytest-playwright
playwright
pytest-playwright
aiohttp
Pillow
